+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+  - edge - refresh set to 10 seconds and remove floating time tolerance
+  - edge detail - re-use status from edge page
   - dashboard detail - issue with refreshing deployment output parameter
 
 ## [2.1.1] - 2019-06.12